+++ src/main/java/webserver/Status.java
@@ -1,4 +1,4 @@
-package common.response;
+package webserver;
 
 public enum Status {
